Workaround missing pkgconfig for zlib on darwin
darwin does not include a pkgconfig for zlib even if it provides the corresponding headers and libraries (see https://github.com/mesonbuild/meson/issues/2654). As a result zlib_dep is not found, and the final build fails with missing symbols. This change is to support locating the same library using the compiler object instead but guarded to do so only on darwin.