Update the 1.1.5 branch to provide a native target and dependency. (#4)
* Update the 1.1.5 branch to provide a native target and dependency.

This is necessary for cross-compilation builds, where we want to link a test program on the build machine only. Without a separate dependency and library that specifies `native: true`, there's no way to use this library with `native: true` targets.

To accomplish this without duplication, we use two dictionaries with a set of machine and compiler info, with one loop handling the build machine and another the host machine.

I also set build_by_default to false for both library targets, because it was causing a cross-compilation build failure in my setup even though I'm not referencing libcmocka, only libcmocka_native.

* Address review feedback