Initial support for expat version 2.2.9
This is based on the 2.2.6 code, with updates for 2.2.9. Once notable
change for the wrap is that libbsd use now defaults to "disabled", just
like it does in the upstream project. I've also added the unit tests,
but I've left the behind a flag, due to having to disable
gnu_symbol_visiblity to build them on platforms with that option.