sanity checks: only run the tests for the current wrap
When that wrap does fallback on some other project, that other
project's tests will often be drawn in too. This is generally wasteful,
since we don't want to test wraps that are already getting tested
themselves.

Occasionally it is also a cause of errors. For example, GLib disables
the tests because they aren't expected to all run, some are flaky. Now
every wrap that *may* fall back on GLib, generally does on Windows, then
runs the GLib tests which fail. Unless the ci_config options for GLib
are copied to a bunch of other wraps.

So, limit the suite being run to only what we care about immediately in
the moment.