fixup! reduce pointless uses of pymod.find_installation
This was supposed to find 'python3', which meson internally guarantees
as sys.executable if it is not otherwise found. Due to a typo, it was
just 'python', which may not exist on silly distros, or may be python2
on other silly distros, and on some *really* silly OSes the only
commands that exist may be `/usr/local/bin/python2.7` and
`/usr/local/bin/python3.8`, which is a bit impossible to find without
relying on sys.executable

The move away from find_installation was supposed to be
behavior-preserving, specifically, it preserves the behavior that
find_installation *also* can detect its own sys.executable.