mpdecimal: de-complexify the configuration stage
Upstream checks for a bunch of things with autotools `AC_*` macros, but
never actually `#include`s the generated config.h or references any of
the `HAVE_` defines. There is no point checking for any of this (unused
configure-time compiler checks are slow), nor writing out an unused
file. Strip this information out.

Some probes that *are* done are only used to figure out which machine
the software is being built for, in order to do other configure.ac /
meson.build logic -- these don't need to be set in configuration_data as
they aren't configured into files, we can just set boolean variables
tracking this information. Using configuration_data is also dangerous
because it may or may not actually be set.

While we are at it, instead of setting unused defines for two possible
bugs:
- glibc mmove
- ipa pure const

add the project arguments which autotools adds.