CI: fix various file reads that don't explicitly use utf-8
Because python doesn't guarantee this, oddly enough (despite being very
onboard the utf-8 train elsewhere), and the default is to use the
platform default, which is... utf-8 on Linux, but the ever-charming
cp1252 on Windows.

Meson requires that meson.build etc. files use the utf-8 encoding, and
by the same token, allows these files to contain non-ASCII contents
(comments are popular). Adding such files to a WrapDB wrap successfully
builds the project, but causes the sanity checks to fail in other ways.