pcre2: Pass libraries as positional argument to pkg.generate()
This fixes an error when using the wrap:
```
subprojects/pcre2-10.40/meson.build:136: DEPRECATION: Library pcre2-8 was passed to the "libraries" keyword argument of a previous call to generate() method instead of first positional argument. Adding libpcre2-8 to "Requires" field, but this is a deprecated behaviour that will change in a future version of Meson. Please report the issue if this warning cannot be avoided in your case.
```

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>