pcre2: Define PCRE2_STATIC when building statically
This fixes the use of `__declspec(dllimport)` and
`__declspec(dllexport)` in the compiled library so that symbols are
exported correctly. Without it, linking against the compiled static
library with VS2017 gives errors like:
```
src_pcre2test.c.obj : error LNK2019: unresolved external symbol __imp_pcre2_config_8 referenced in function process_command
```

See https://gitlab.gnome.org/rybalkin/glib/-/jobs/2119305 for an example
of the broken (previous) version of this, and
https://gitlab.gnome.org/rybalkin/glib/-/jobs/2119406 for an example of
a fixed (this) version.

Adding the new define to `declare_dependency()` is not strictly
necessary, but makes it easier to correctly use the built static library
from another project.

This releases version 10.40-3 of the pcre2 wrap.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>