sanity checks: verify that a wrap is actually unsupported
Historically, we have a setting "linux_only" for wraps which are
software that targets Linux APIs and doesn't intend to compile
elsewhere. After adding more CI coverage, we changed this to "build_on"
as an operating system key. The intention is that we don't need to fail
CI for software that is performing within its expected parameters.

Unfortunately, this has been repeatedly misused for wraps that are
currently broken and need to be fixed, which defeats the point of having
CI at all. And these PRs get merged when they shouldn't be despite
review objections.

Stop skipping these wraps. We should verify that they actually fail, and
why. This has a couple benefits:

- if a wrap is disabled because it is missing dependencies, that's not
  ideal, but tolerable. Still, disabling the CI for it means that if the
  dependency becomes satisfiable, we still skip CI. Now, we can raise an
  error that it is satisfiable and can configure properly, and needs to
  be unblocked in ci_config.json

- if a wrap is disabled because the software isn't coded for a platform,
  mysterious configure failures are a bad user experience, it's far
  better to error out with `error('foobar is not supported on Windows')`

- if a wrap is disabled because, say, it compiles successfully on
  Windows but produces broken DLLS that have no symbol table or implib,
  then something went badly wrong and that's not a good reason to
  disable it