json-c: fix SIZEOF_SSIZE_T
On my setup SIZEOF_SSIZE_T ends up being negative:

    #define SIZEOF_SSIZE_T -1

This happens because <stdint.h> is not enough to pull the
definition of ssize_t.

Instead, use <sys/types.h> as per the POSIX spec [1].

While at it, also fix the same issue for size_t, which was pulled
in by chance. Switch to <stddef.h> which is ISO C.

[1]: https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/sys_types.h.html