CI: do not run scheduled jobs in forks
Enabling CI in forks is reasonable because people might want to enable
CI in order to test commits before submitting them as PRs. It's a bit
less reasonable to run a weekly job no matter what -- this is intended
primarily for upstream WrapDB usage to detect and investigate
regressions, and won't provide meaningful results distinct from the
upstream repository if run on forks. All it does is waste time and
resources.

Like mesonbuild/meson's images.yml, this also allows running even on
forks, if started for a reason other than automatic scheduling. In the
case of Meson, it's still interesting to test whether updates to the
image generation routines work, and in the case of the WrapDB, someone
might want to manually test a fork for the same reason people with write
permissions to the WrapDB would want to test it.

In both cases, doing so is fundamentally opt-in.