Add support for choosing which Meson repo and branch to test with
Every so often there are PRs which change the way some of Meson's core
inner workings operate. Although they may have passed the CI, Meson's
test suite isn't always guaranteed to be complete. By allowing a PR
author the ability to test against all of wrapdb, they and the Meson
maintainers can be more assured that the PR is correct.