ci: Do not fail build-all job for expected failures
Some wraps (e.g. Wayland) fail on some platforms as expected but without
an explicit unsupported error message. We consider that to be a failure
for PRs, but for the build-all job it is not a problem.

Wraps that were expected to fail but suddenly succeed is still an issue
that should fail the build-all job.

Fixes: #906