pcre2: fix bad usage of cmake-formatted config files
pcre2.h.in isn't even in that format to begin with so it was weird to
use it.

config-cmake.h.in was checking for a lot of stuff it never used (and in
one case, doing so twice), and missed checking for something that is
used (this is a bug in the cmake definition upstream). It's not
especially portable across meson versions, and the cmake template is
basically just a list of defines anyway, so it's extremely trivial to
generate this using meson's header dumping ability.