sdl2: provide sdl2main and sdl2_test dependencies
This better reflects how sdl2main should be used and actually works when a library (e.g. sdl2_image) links to sdl2 in a project that builds with `with_main=true`.

Mark `with_main` deprecated and default it to false - this is a sensible default now that linking to sdl2main the right way is simple.

sdl2_test is used for testing by some other projects, e.g. sdl2_image - which motivated me to do this change in the first place.

Additionally, bump required meson version to 0.60 and replace deprecated functions with alternatives.