zlib: fix broken "solo" option
Mostly because upstream has a couple quirks here. The really annoying
one is that the highly custom configure script modifies zconf.h in-place
with sed, which is quite ouchy. Really, Z_SOLO needs to be defined for
all uses of zconf.h -- this broke building example.c.

The other issue is that the Makefile specifically excludes a number of
source files for libz when solo is active, which should not be compiled.
We didn't actually try to do that, so those source files failed to
build. Fix this by not building them.

Fixes #986