liburing: override "liburing" dependency name
The pkg-config name of liburing is "liburing", so ideally it would be
possible to use just

  dependency('liburing')

to find either the system provided liburing or fall back to the subproject.
However, that does not work since the liburing wrap project does not
override or provide the "liburing" dependency name, so one has to use

  dependency('liburing', fallback: ['liburing', 'uring'])

or

  dependency('liburing', 'uring')

This change makes it possible to use

  dependency('liburing')

with both system (pkg-config) and the wrapdb provided liburing.