pkgconf: use correct download url with different tarball
The official download instructions have always been to download from a
files hosting site -- the distfiles subdomain of dereferenced.org /
ariadne.space.

Instead this used some autogenerated tarball from gitea, which is not
very reliable IMO, and has the additional disadvantage of not putting
the tag into the extracted directory name so every version clobbers each
other.

Follow the very clearly worded directions in the pkgconf readme:

> Release tarballs are available at https://distfiles.ariadne.space/pkgconf/. Please only use the tarballs from distfiles.ariadne.space.

This is the same tarball, same checksums, that other redistributors of
pkgconf use to build with.