emilk-loguru: Fix formatting + edit description
Some descriptions of options contained superfluous whitespace. This
commit removes it.

Descriptions for loguru_debug_logging and loguru_debug_checks say that
these options default to the opposite of NDEBUG. This is true in general
but meson fully overrides these options so that they are enabled by default
independently of NDEBUG (or b_ndebug).