Add libsass 3.6.4
This is a direct copy of https://github.com/lazka/libsass/tree/meson
which I've created for gtk back in the days.

There is an open PR for adding meson support upstream in
https://github.com/sass/libsass/pull/3073 but the project is considered deprecated
and hasn't seen a commit in over a year, so there is not much hope for progress there.

There is a newer 3.6.5 release out, but I wanted to copy things as is from my fork first
and look at updating it later.