Add sassc 3.6.2 (#1185)
This is a direct copy of https://github.com/lazka/sassc/tree/meson
which I've created for gtk back in the days.

This project is in a similar situation as #1153. There exists
an upstream PR for adding meson (https://github.com/sass/sassc/pull/269),
but the project is unmaintained, so unlikely to see progress there.