openssl: Make it build with MSVC
MSVC needs a way of knowing which symbols are supposed to be exported,
otherwise it does not create a DLL. This means the wrap was previously
unusable as a shared library.

The OpenSSL build process does that by passing a `.def` file to the
linker. This reimplements OpenSSL's `util/mkdef.pl` script to make the
meson build do the same as the normal build process.

Fixes #181.