nghttp2: Fix use with `[wrap-git]` by outputting `nghttp2ver.h`
When using nghttp2 with `[wrap-git]` you currently get a
compilation error:

    nghttp2.h:55:10: fatal error: nghttp2/nghttp2ver.h: No such file or directory

It works with `[wrap-file]`, because the tarballs comes with
`nghttp2ver.h`. But it is not versioned in git and is thus
missing `[wrap-git]`.

Fix what appears to be a typo so that meson can generate it
also with a git checkout.