sanity_checks: also disable builds by CPU family
In some cases the OS is not enough to determine if a build will succeed
or fail, the CPU architecture (family) may also change the result.
This commit extends the sanity_checks to consult the host CPU family
(as determined by Meson), if either the OS or family is mapped to
`false` in ci_config.json an error is expected. The composite key
`{system}-{cpu_family}` can be used to expect success (or failure) for
particular combinations otherwise marked for failure (or success).