cmocka: drop library name_prefix on Windows
The current version of cmocka has source file src/cmocka.def, which
provides the list of library symbols for MSVC builds. The library name
in that file is simply "cmocka", which results in projects using the
library trying to link to "cmocka.dll". When cross compiling with Meson
and MinGW-w64 though, the actually built library file is libcmocka.dll.

For Windows builds, set library name_prefix to empty string, so that
Meson builds cmocka.dll, as expected by binaries linked to the library.
Keep the default prefix for other operating systems.

Fixes: https://github.com/mesonbuild/wrapdb/issues/1612
Suggested-by: Rosen Penev <rosenp@gmail.com>
Signed-off-by: Grzegorz Szymaszek <gszymaszek@short.pl>