sqlite3: remove -D_GNU_SOURCE from exported dependency object
We weren't setting this when compiling sqlite3.c, but only in the
compile_args passed to the superproject via sqlite3_dep.  sqlite3.h
doesn't care about this macro, and we shouldn't force the superproject to
set a define with global consequences, so remove the flag.

In any event, we don't need _GNU_SOURCE when compiling sqlite3.c because
it already defines that macro, so it's not clear why this was here in the
first place.

This effectively reverts commit a48d9345490197f5ce054ef9cb2d8b00ade3b925.