luajit: Detect failure due to unsupported architecture
This is needed communicate to WrapDB's CI when it is normal for
configuration to fail.

To my knowledge there's no good way to obtain the compiler's error
message in meson, so we cannot distinguish errors due to unsupported
architectures from errors due to otherwise incorrect configurations,
so this does have a slight risk of false positives in marking failures
as "intended". The alternative (apart from making an additional python
script) would be manually checking host_machine.cpu_family, but this
would need to be manually updated when luajit starts supporting more
architectures.