spdlog wrap: Added a meson option to specify whether or not to install headers.
Installing headers and pkg-config files is not always wanted, especially as spdlog is naturally used as a subproject. After compilation, one does not need the header files anymore, except when compiling spdlog as a shared library.

This patch adds an option to disable installation, but does not change the current default behaviour.
The current CMAKE build file of spdlog currently contains this option as well, so it only makes sense to replicate it.