freeglut: fix broken wayland build
The most recent commit, 8472c965 ("freeglut: add missing dependencies")
broke when this was built for wayland, in two ways:

1. It dropped the xkbcommon dep, which is required on Wayland as well
2. It started passing x11_dep to the XParseGeometry-detection, but
   x11_dep isn't set in that code-path.

The former was fixed simply by adding it back. The latter was fixed by
dropping the x11_dep in the Windows code-path, and rather passing all of
freeglut_deps to the detection logic.