sanity_checks: auto-disable fatal warnings if CMake warning is in logs
New Meson releases sometimes add dubious warnings that break existing
packages built with --fatal-meson-warnings, which is the WrapDB default.
Add a mechanism to detect when setup fails on a warning that matches a
regex and automatically rerun setup without --fatal-meson-warnings.  This
is a large hammer, but allows us to continue making forward progress in
this repo while fixes percolate into Meson.

For now, match the new warning in 1.7.1+ that triggers if CMake is used
to detect a dependency when the Find*.cmake file is shipped with CMake
itself and the dependency is not available.  This shouldn't be fatal, and
in fact it triggers even for optional dependencies.