sanity_checks: in PR CI, avoid tag check racing with tag creation
test_releases_json validates that all existing tags are mentioned in
releases.json.  This ensures PRs don't mangle or remove old releases.json
entries, but poses the potential for races in CI.  We don't want PR CI to
fail because of a newly-created tag the PR branch doesn't know about.

If a PR is created from a stale branch, this doesn't cause a problem.
actions/checkout checks out a synthetic merge commit against the current
master branch, so the releases.json seen by sanity_checks.py includes any
newly-tagged versions even if the PR branch does not.  Likewise, if a tag
is created after actions/checkout runs, the CI job doesn't see the new
tag, so there's no problem.

However, if a tag is created *after* the synthetic merge commit but
*before* a job runs actions/checkout, that job's test_releases_json will
fail.  This is annoying because merging a PR breaks any pending CI for
other PRs.  Each PR today runs 15 CI jobs, we only get 20 concurrent
slots, we have some slow emulated-CPU jobs, and wrapdb-bot tends to submit
multiple PRs at once, so it's not uncommon to have pending jobs in the
queue.

Avoid the race by ignoring tags that aren't reachable from HEAD, which for
PR CI is the synthetic merge commit and for branch CI is the tip of the
branch.  This also prevents branch CI from failing if the branch is stale,
though that's mainly relevant when running CI in a fork.