tools/versions: always allow autoupdate --revision
--revision currently a) refuses to bump the revision of projects with
upstream Meson support or whose current version is not tracked by Anitya
and b) bumps the version instead of the revision if a new version is
available.  Both behaviors are based on faulty assumptions.  We might
plausibly want to bump a wrap's revision after changing the wrap file,
regardless of the project's status in other respects.  And it's reasonable
to want to fix something in a Meson port without taking responsibility for
updating it to the latest version.

Change the semantics of --revision to always bump the revisions of the
specified wraps, even if --port is not specified.  If no wraps are
specified, bump all wraps for consistency.

Rework the help text to describe the various things autoupdate can do,
rather than only emphasizing projects with upstream Meson support.