Add package name to source_filename for wraps that don't have it
A number of wraps use a GitHub archive URL in source_url.  The filename
component of these URLs is <tagname>.tar.gz, so the wraps naturally
configure <tagname>.tar.gz as the source_filename.  But the
source_filename is arbitrary, so they're not constrained to do that, and
there are a couple reasons not to:

- subprojects/packagecache tends to accumulate old versions of subproject
  tarballs (in a wrap-heavy workflow, anyway) and it's not obvious whether
  a file called "v1.7.18.tar.gz" is still needed.

- source_filename is a global namespace.  If two wraps happen to use the
  same source_filename, Meson will fail on a hash mismatch when it tries
  to unpack the second one:

    meson.build:10:8: ERROR: Incorrect hash for source:
     9a93b2b7dfdac77ceba5a558a580e74667dd6fede4585b91eefb60f03b72df23 expected
     3aa806844a03442c00769b83e99970be70fbef03735ff898f4811dd03b9f5ee5 actual.

Change the source_filename in all such wraps to the value of `directory`
plus the original file extension(s).