freeglut: require GL/glu.h
Freeglut requires GL/glu.h, which isn't always installed. Make sure we
check for it, so we can get a meaningful error message early on instead
of just failing to compile.

Because not all platforms has glu as a dependency, we also need to fall
back to searching for the library. This search is copied from Mesa's
demos project.