backward-cpp: fix helper library selection
We're only supposed to select one stack walker (at most) and one stack
detail extractor (at most).  Instead we were linking with everything we
found, enabling all extractors we found, and failing to enable libunwind
even if we found it.  Fix this, and add options so the user can avoid an
available high-priority extractor in favor of a lower-priority one.