libjpeg-turbo: fix ARM Neon intrinsics on GCC
If the neon-intrinsics option was set to auto, we always disabled it.
When disabled, we still added the define claiming it was enabled.  As a
result, if the option was not set to enabled, many tests failed on aarch64
GCC with errors like:

    Testing 8-bit precision
    RGB Top-Down  -> 4:4:4 Q100 ... Done.
      Result in test_enc8_RGB_TD_444_Q100.jpg
    JPEG -> RGB Top-Down  2/1 ... TurboJPEG ERROR:
    Corrupt JPEG data: premature end of data segment
    JPEG -> RGB Top-Down  15/8 ... TurboJPEG ERROR:
    Corrupt JPEG data: premature end of data segment

Presumably these reflected real problems in the binary.  CI only had
aarch64 Clang until recently, and on Clang with Meson >= 1.2 this didn't
happen: we force-enable the option because the partial Neon intrinsics
check fails.

Fixes: 1b0dbb4532 ("libjpeg-turbo: add Arm support")