Require valid license declaration for updated projects in this repo
WrapDB includes a lot of projects with missing license declarations or
invalid SPDX expressions.  Projects with upstream Meson configs have a
longer fix cycle, so it'd be too painful to fail CI on invalid licenses,
but for downstream Meson configs we can certainly require fixes when wraps
are updated.

meson-info doesn't mention the licenses of subprojects, but we can extract
that information from the rewriter, if it is working.  (Extraction fails
if the Meson requests a compiler that isn't installed.)

Use the license-expression library, if installed, to validate the SPDX
expression.  Install it in every sanity-check job since we don't know
which jobs will successfully run the rewriter.

Of course, CI can't force the license declaration to be _correct_, but
then again, upstream may not have it right either.