pcre2: improve function checks
Several functions are only needed if jit_sealloc is enabled, so
check them only on that case.

The code for jit_sealloc doesn't build if `mkostemp()` is unavailable
and a condition was already in place to check for that, expand it to
also include `secure_getenv()` which is used in the same way.