openssl: update to 3.0.21
This also enables capturing .asm and .rc files, and updates the existing
logic to support MSVC 32 and 64-bits x86. This was already available for
the OpenSSL 1.x version of my fork, but I had never upstreamed its
support.

I've also bumped the version to 0.64 to enable the Nasm language
support.