Work around Meson bug ignoring c_std/cpp_std in subprojects
On Meson versions affected (or expected to be affected) by
mesonbuild/meson#14939, add C and C++ languages to the root project so
subprojects will respect c_std and cpp_std.

We can adjust the version range if needed once the bug is fixed.  There's
probably some value in continuing to carry this check for a while after
the fix is released, since developers might have older Meson versions on
their machines.