sanity_checks: treat releases.json provides as canonical
Wraps implicitly provide the wrap file name, and a subproject may or may
not choose to add an override with that name.  Currently we treat the wrap
file provides as canonical and check releases.json against them, so if no
explicit self-named provide is given, we can't know whether the subproject
intends the implicit one.  To prepare to support the implicit case, treat
releases.json as canonical instead, and check the wrap file against it.

Also refactor the check logic a bit.

No functionality change intended, except that the root meson.build will
now check for deps and progs in sorted order.