Use latest Meson prerelease for 'meson format'
Formatting rules change over time, and CI uses the latest Meson
prerelease, so we should match that version on developers' systems.  If
not running in CI, install Meson into a virtualenv that we update at most
once per day, and use it for format checks in sanity_checks.py and
format.py.