create_release: move source_fallback_url key before patch_*
If source_fallback_url doesn't already exist in the wrap file, we add it
at the end of the [wrap-file] section, which means we've been putting it
after the patch_* keys created by create_patch_zip().

create_source_fallback() and create_patch_zip() are independent of each
other, so reverse their order to keep source_fallback_url next to the
other source_* keys.