ci: add IBM-hosted ppc64le and s390x runners
They're only available in mesonbuild/wrapdb, not in forks.  As a result
the emulated ppc64le job is potentially still useful, but it would likely
bitrot since it wouldn't run in the main repo.  Drop that job.

Self-hosted runners don't count toward the concurrent job limit so this
actually improves our parallelism.