create_release: upload to draft release, then publish
We've been creating a public release and then uploading files into it.
However, if we crash halfway through, we leave behind a Git tag and an
unusable release without a wrap file, since that's the last thing to be
uploaded.  Subsequent runs see the tag and don't retry.

(Also, if we enable GitHub's immutable releases feature in the future,
GitHub won't allow us to upload to a published release.)

Instead, create the release in draft form, upload to it, and then publish
the release.  GitHub doesn't create the tag until we publish, so
newly-created tags will reflect only releases that have actually
completed.

Fixes: #2224