create_release: delete stale draft from previous run
Now that we don't tag releases until they're complete, we retry failed
releases on subsequent runs.  However, if we locate an existing draft
release and any assets have been uploaded to it, GitHub refuses to let us
upload those again, so the retry fails.  Instead of manually clearing out
existing assets, just delete existing drafts with the intended tag name.

If the existing release is not a draft, we've somehow failed to clone a
complete set of tags for the repo, which should be impossible if
release.yml is correctly imposing concurrency control.  Fail in that case.