create_release: tag the checked-out commit, not upstream HEAD
create_release always tags the repo's current HEAD commit, not the commit
checked out by the release job.  Thus, if two PRs that modify the same
wrap land in quick succession, we could release the contents of the first
PR under a tag attached to the second one.  If both PRs include identical
release bumps (as seems likely), merge will elide the second one, so the
second PR will not cause an additional release with the intended contents
but a manual investigation will find the tag in the expected place.

Pass the commit hash of the current checkout during release creation,
ensuring the correct commit is tagged.