create_release: add patch_fallback_url
patch_url points to a redirect on wrapdb.mesonbuild.com, which is
convenient if we ever want to add additional mirrors or if we have to move
off of GitHub.  However, it also means wraps will fail to build if
wrapdb.mesonbuild.com goes down, unless the patch is already cached
locally.  Avoid this by adding a patch_fallback_url pointing directly to
the GitHub download.