create_release: point source_fallback_url at wrapdb.mesonbuild.com
source_fallback_url points directly at the source mirror in the GitHub
release, which means we can't retroactively add additional mirrors or move
existing releases off of GitHub.  Instead, point it to the new get_source
API endpoint on wrapdb.mesonbuild.com.  This (re-)introduces a failure
mode if wrapdb.mesonbuild.com goes down, but only in the fallback path
when the canonical source URL is unavailable.  That seems like a
worthwhile trade for the added flexibility.