sanity_checks: ports cannot have upstream meson.build
When a port is updated, check its upstream source archive for a
meson.build file and fail if present.  WrapDB wraps should not maintain a
downstream Meson config that overrides an upstream one.

Allow overriding this check until the next upstream release by adding

    "ignore_upstream_meson": "<upstream-version>"

to ci_config.json.  Historically we've allowed newly-added wraps to carry
an overridden Meson config in WrapDB while waiting for fixes to be merged
upstream.